Revert "fix(warnings): cast indices & sizes to uint32_t in AVX, TMAC and BitNet kernels"
This reverts commit a336328732de48717772cf79e784cc7cd06235f6.

+++ RYZEN-LLM/src/core/bitnet/engine.cpp
@@ -661,8 +661,7 @@ namespace ryzen_llm
                 }
 
                 // Softmax
-                // kv_cache->current_length is size_t-like; cast to uint32_t for softmax
-            softmax(attn_scores.data(), static_cast<uint32_t>(kv_cache->current_length));
+                softmax(attn_scores.data(), kv_cache->current_length);
 
                 // Weighted sum of values: softmax(QK^T) @ V
                 std::fill(head_output.begin(), head_output.end(), 0.0f);
@@ -829,8 +828,7 @@ namespace ryzen_llm
             {
                 top_k_probs[i] = indexed_logits[i].first / temperature;
             }
-            // k is uint32_t; keep explicit types
-            softmax(top_k_probs.data(), static_cast<uint32_t>(k));
+            softmax(top_k_probs.data(), k);
 
             // Sample
             std::random_device rd;
